ESP32_LVGL/LVGL8/5_System_Monitor/PythonScriptForSysMonitor/requirements.txt
psutil
pyserial